move agent config under /internal/agent/config
