+++ internal/agent/config/config_test.go
@@ -1,4 +1,4 @@
-package agent
+package config
 
 import (
 	"flag"
